Partial build fix for WSL, disabling seeder
Summary:
On the Windows Subsystem for Linux (WSL), the compilation
fails because of `bitcoin_seeder`. As this app is not a
must-have, the build instructions are updated with the
proper hint.

Still on WSL, a fix from Bitcoin Core is backported in order
to get the tests to compile. See:
https://github.com/meritlabs/merit/commit/c1e291a6ac7ea7ae57697541e308813a87622325

Bitcoin-ABC/bitcoin-abc#190
Bitcoin-ABC/bitcoin-abc#228

Test Plan: Compiles and green tests on Win10 x64 under Ubuntu subsystem.

Reviewers: O1 Bitcoin ABC!, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D1602